feat: implémentation des update / delete

+++ app/src/main/java/ch/heigvd/iict/and/rest/ContactsRepository.kt
@@ -3,6 +3,7 @@ package ch.heigvd.iict.and.rest
 import ch.heigvd.iict.and.rest.database.ContactsDao
 import ch.heigvd.iict.and.rest.database.RemoteSyncManager
 import ch.heigvd.iict.and.rest.models.Contact
+import ch.heigvd.iict.and.rest.models.Status
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -21,19 +22,30 @@ class ContactsRepository(private val contactsDao: ContactsDao, private val remot
 
     fun update(contact: Contact) {
         CoroutineScope(Dispatchers.IO).launch {
+            contact.status = Status.MODIFIED
+            contactsDao.update(contact)
+            remoteSyncManager.editContact(contact)
             contactsDao.update(contact)
         }
     }
 
     fun create(contact: Contact) {
         CoroutineScope(Dispatchers.IO).launch {
+            contact.status = Status.NEW
             contactsDao.insert(contact)
+            remoteSyncManager.registerContact(contact)
+            contactsDao.update(contact)
         }
     }
 
     fun delete(contact: Contact) {
         CoroutineScope(Dispatchers.IO).launch {
-            contactsDao.delete(contact)
+            contact.status = Status.DELETED
+            contactsDao.update(contact)
+            remoteSyncManager.deleteContact(contact)
+            if (contact.status == Status.OK) {
+                contactsDao.delete(contact)
+            }
         }
     }
